requirements.txt
requests==2.32.3
Django==4.2.20
django-phone-field==1.8.1
django-phonenumber-field==8.1.0
whitenoise==6.9.0
geoip2==5.1.0
phonenumbers==8.13.33
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0